Update Readme.md
This file has been edited by a Computer Science student in a Technical Writing course. The students are tasked in finding online real-world documentation to edit. Full marks are achieved if the original author of the work accepts the changes; in this case a Pull Request.
Please review and accept the Pull Request and thank you so much!